RDM-4324 Refactor compound case fields sorting logic into CompoundFieldOrderService. Clean up method names.
