+++ src/test/java/uk/gov/hmcts/ccd/v2/internal/controller/UIStartTriggerControllerTest.java
@@ -75,7 +75,6 @@ class UIStartTriggerControllerTest {
                     .withDisplayContext("MANDATORY")
                     .withLabel(FIELD_LABEL)
                     .withHintText(FIELD_HINT_TEXT)
-                    .withOrder(1)
                     .withShowCondition(null)
                     .build()))
             .build())
